feat: ensure VideoController returns 404 if a video return not found on delete

+++ backend/app/controllers/VideoController.ts
@@ -78,6 +78,10 @@ export default class VideoController {
     try {
       const { uuid } = request.params()
 
+      const video = await Video.findBy('uuid', uuid)
+      if (!video) {
+        return notFound()
+      }
       await Video.query().where('uuid', uuid).delete()
 
       return noContent()
